+++ RealmSwift/Sync.swift
@@ -150,7 +150,7 @@ public struct Credential {
         self.userInfo = userInfo
     }
 
-    private init(_ credential: RLMSyncCredential) {
+    internal init(_ credential: RLMSyncCredential) {
         self.token = credential.token
         self.provider = credential.provider
         self.userInfo = credential.userInfo
@@ -185,7 +185,7 @@ public struct Credential {
 }
 
 extension RLMSyncCredential {
-    fileprivate convenience init(_ credential: Credential) {
+    internal convenience init(_ credential: Credential) {
         self.init(customToken: credential.token, provider: credential.provider, userInfo: credential.userInfo)
     }
 }